Using Hash Table Instead of Array in Breakpoint List
While debugging a program with a big number of breakpoints, it's
necessary a better structure to hold the list instead of a simple
array list.

This commit introduces a hashtable data structure and uses it for
the breakpoint list, which decreases the time spent in
'bp_findbreakpoint' by a factor of 20 and decreases the overall
time by something around ~30% lesser.