Enumerations Support
This commit introduces support for enumerations in C. Support
for enums took a long time because I thought there was some
way to discover the signedness of the integer type chosen by
the compiler. GCC adds the DW_AT_type field to the enum
definition in the DWARF structure. However, Clang does not
emit this field and therefore we cannot rely on its existence
to discover the signal of the type.

There would be the possibility of taking a 'mixed' approach
that checks for the existence of the field and, if any,
analyzes the base type signal. However, to keep it simple,
I will always assume enum as a signed type.