Bug Fix: Wrong Index If Multiples Changes
When offmemcmp detects more than one change in a vector/matrix, it
returns the byte_offset accordingly to the pointers passed as a
parameter. It turns out that these pointers can be incremented as
new changes are found and the byte_offset alone does not match the
real byte_offset in which the change occurred.

Therefore, this commit corrects this taking into account the new
pointer address and the base address, and thus, correctly
calculating the indexes, either with one or more changes.